+++ README.md
@@ -12,8 +12,6 @@ Controls are simple:
 * S: Down
 * D: Right
 * Q: Quit
-* 1: Toggle Enemy Movement
-* 2: Toggle Enemy Spawning
 * Escape: Pause
 * Left-Click: Fire
