--- font_conv.zig
@@ -1,130 +1,0 @@
-pub export const FONT_A: u64 = @as(u64, @bitCast(@as(c_long, 4352278784536462275)));
-pub export const FONT_B: u64 = 18213616809356280828;
-pub export const FONT_C: u64 = @as(u64, @bitCast(@as(c_long, 4593883554018017343)));
-pub export const FONT_D: u64 = 17351458709922106608;
-pub export const FONT_E: u64 = 18446674804472872959;
-pub export const FONT_F: u64 = 18446674804472856768;
-pub export const FONT_G: u64 = @as(u64, @bitCast(@as(c_long, 9223302625347486718)));
-pub export const FONT_H: u64 = 14106333962129621955;
-pub export const FONT_I: u64 = 18446489090495414271;
-pub export const FONT_J: u64 = @as(u64, @bitCast(@as(c_long, 217020518514230268)));
-pub export const FONT_K: u64 = 14397120638141123783;
-pub export const FONT_L: u64 = 13889313184910721279;
-pub export const FONT_M: u64 = 14116532876766069699;
-pub export const FONT_N: u64 = 14115393885800484807;
-pub export const FONT_O: u64 = @as(u64, @bitCast(@as(c_long, 9151248213410578302)));
-pub export const FONT_P: u64 = 18374620251275837632;
-pub export const FONT_Q: u64 = @as(u64, @bitCast(@as(c_long, 9151248213414477315)));
-pub export const FONT_R: u64 = 18357732005822123207;
-pub export const FONT_S: u64 = @as(u64, @bitCast(@as(c_long, 9205569002326393854)));
-pub export const FONT_T: u64 = 18446489090495354904;
-pub export const FONT_U: u64 = 14106333703424951166;
-pub export const FONT_V: u64 = 14106333302423829528;
-pub export const FONT_W: u64 = 14106333703829186499;
-pub export const FONT_X: u64 = 14106230866751964099;
-pub export const FONT_Y: u64 = 14106230866142828568;
-pub export const FONT_Z: u64 = 18446470426191265791;
-pub export const FONT_0: u64 = 18411139144890810879;
-pub export const FONT_1: u64 = @as(u64, @bitCast(@as(c_long, 1736164148113840152)));
-pub export const FONT_2: u64 = @as(u64, @bitCast(@as(c_long, 9116136719743213823)));
-pub export const FONT_3: u64 = 18159077807387378428;
-pub export const FONT_4: u64 = 9331882835121864961;
-pub export const FONT_5: u64 = 18410857109412446719;
-pub export const FONT_6: u64 = @as(u64, @bitCast(@as(c_long, 8971311748128866940)));
-pub export const FONT_7: u64 = 18374970170986733632;
-pub export const FONT_8: u64 = @as(u64, @bitCast(@as(c_long, 9115709501113205118)));
-pub export const FONT_9: u64 = @as(u64, @bitCast(@as(c_long, 9115709503252267265)));
-pub export const FONT_COLON: u64 = @as(u64, @bitCast(@as(c_long, 6781788124348416)));
-pub export const FONT_CHAR_TABLE: [256]u64 = [91]u64{
-    @import("std").mem.zeroes(u64),
-    @import("std").mem.zeroes(u64),
-    @import("std").mem.zeroes(u64),
-    @import("std").mem.zeroes(u64),
-    @import("std").mem.zeroes(u64),
-    @import("std").mem.zeroes(u64),
-    @import("std").mem.zeroes(u64),
-    @import("std").mem.zeroes(u64),
-    @import("std").mem.zeroes(u64),
-    @import("std").mem.zeroes(u64),
-    @import("std").mem.zeroes(u64),
-    @import("std").mem.zeroes(u64),
-    @import("std").mem.zeroes(u64),
-    @import("std").mem.zeroes(u64),
-    @import("std").mem.zeroes(u64),
-    @import("std").mem.zeroes(u64),
-    @import("std").mem.zeroes(u64),
-    @import("std").mem.zeroes(u64),
-    @import("std").mem.zeroes(u64),
-    @import("std").mem.zeroes(u64),
-    @import("std").mem.zeroes(u64),
-    @import("std").mem.zeroes(u64),
-    @import("std").mem.zeroes(u64),
-    @import("std").mem.zeroes(u64),
-    @import("std").mem.zeroes(u64),
-    @import("std").mem.zeroes(u64),
-    @import("std").mem.zeroes(u64),
-    @import("std").mem.zeroes(u64),
-    @import("std").mem.zeroes(u64),
-    @import("std").mem.zeroes(u64),
-    @import("std").mem.zeroes(u64),
-    @import("std").mem.zeroes(u64),
-    @import("std").mem.zeroes(u64),
-    @import("std").mem.zeroes(u64),
-    @import("std").mem.zeroes(u64),
-    @import("std").mem.zeroes(u64),
-    @import("std").mem.zeroes(u64),
-    @import("std").mem.zeroes(u64),
-    @import("std").mem.zeroes(u64),
-    @import("std").mem.zeroes(u64),
-    @import("std").mem.zeroes(u64),
-    @import("std").mem.zeroes(u64),
-    @import("std").mem.zeroes(u64),
-    @import("std").mem.zeroes(u64),
-    @import("std").mem.zeroes(u64),
-    @import("std").mem.zeroes(u64),
-    @import("std").mem.zeroes(u64),
-    @import("std").mem.zeroes(u64),
-    FONT_0,
-    FONT_1,
-    FONT_2,
-    FONT_3,
-    FONT_4,
-    FONT_5,
-    FONT_6,
-    FONT_7,
-    FONT_8,
-    FONT_9,
-    FONT_COLON,
-    @import("std").mem.zeroes(u64),
-    @import("std").mem.zeroes(u64),
-    @import("std").mem.zeroes(u64),
-    @import("std").mem.zeroes(u64),
-    @import("std").mem.zeroes(u64),
-    @import("std").mem.zeroes(u64),
-    FONT_A,
-    FONT_B,
-    FONT_C,
-    FONT_D,
-    FONT_E,
-    FONT_F,
-    FONT_G,
-    FONT_H,
-    FONT_I,
-    FONT_J,
-    FONT_K,
-    FONT_L,
-    FONT_M,
-    FONT_N,
-    FONT_O,
-    FONT_P,
-    FONT_Q,
-    FONT_R,
-    FONT_S,
-    FONT_T,
-    FONT_U,
-    FONT_V,
-    FONT_W,
-    FONT_X,
-    FONT_Y,
-    FONT_Z,
-} ++ [1]u64{@import("std").mem.zeroes(u64)} ** 165;
