--- font_processed.h
@@ -1,399 +1,0 @@
-# 1 "font.h"
-# 1 "<built-in>" 1
-# 1 "<built-in>" 3
-# 390 "<built-in>" 3
-# 1 "<command line>" 1
-# 1 "<built-in>" 2
-# 1 "font.h" 2
-
-
-
-
-# 1 "/usr/lib/clang/18/include/stdint.h" 1 3
-# 52 "/usr/lib/clang/18/include/stdint.h" 3
-# 1 "/usr/include/stdint.h" 1 3 4
-# 26 "/usr/include/stdint.h" 3 4
-# 1 "/usr/include/bits/libc-header-start.h" 1 3 4
-# 33 "/usr/include/bits/libc-header-start.h" 3 4
-# 1 "/usr/include/features.h" 1 3 4
-# 394 "/usr/include/features.h" 3 4
-# 1 "/usr/include/features-time64.h" 1 3 4
-# 20 "/usr/include/features-time64.h" 3 4
-# 1 "/usr/include/bits/wordsize.h" 1 3 4
-# 21 "/usr/include/features-time64.h" 2 3 4
-# 1 "/usr/include/bits/timesize.h" 1 3 4
-# 19 "/usr/include/bits/timesize.h" 3 4
-# 1 "/usr/include/bits/wordsize.h" 1 3 4
-# 20 "/usr/include/bits/timesize.h" 2 3 4
-# 22 "/usr/include/features-time64.h" 2 3 4
-# 395 "/usr/include/features.h" 2 3 4
-# 481 "/usr/include/features.h" 3 4
-# 1 "/usr/include/stdc-predef.h" 1 3 4
-# 482 "/usr/include/features.h" 2 3 4
-# 503 "/usr/include/features.h" 3 4
-# 1 "/usr/include/sys/cdefs.h" 1 3 4
-# 576 "/usr/include/sys/cdefs.h" 3 4
-# 1 "/usr/include/bits/wordsize.h" 1 3 4
-# 577 "/usr/include/sys/cdefs.h" 2 3 4
-# 1 "/usr/include/bits/long-double.h" 1 3 4
-# 578 "/usr/include/sys/cdefs.h" 2 3 4
-# 504 "/usr/include/features.h" 2 3 4
-# 527 "/usr/include/features.h" 3 4
-# 1 "/usr/include/gnu/stubs.h" 1 3 4
-# 10 "/usr/include/gnu/stubs.h" 3 4
-# 1 "/usr/include/gnu/stubs-64.h" 1 3 4
-# 11 "/usr/include/gnu/stubs.h" 2 3 4
-# 528 "/usr/include/features.h" 2 3 4
-# 34 "/usr/include/bits/libc-header-start.h" 2 3 4
-# 27 "/usr/include/stdint.h" 2 3 4
-# 1 "/usr/include/bits/types.h" 1 3 4
-# 27 "/usr/include/bits/types.h" 3 4
-# 1 "/usr/include/bits/wordsize.h" 1 3 4
-# 28 "/usr/include/bits/types.h" 2 3 4
-# 1 "/usr/include/bits/timesize.h" 1 3 4
-# 19 "/usr/include/bits/timesize.h" 3 4
-# 1 "/usr/include/bits/wordsize.h" 1 3 4
-# 20 "/usr/include/bits/timesize.h" 2 3 4
-# 29 "/usr/include/bits/types.h" 2 3 4
-
-
-typedef unsigned char __u_char;
-typedef unsigned short int __u_short;
-typedef unsigned int __u_int;
-typedef unsigned long int __u_long;
-
-
-typedef signed char __int8_t;
-typedef unsigned char __uint8_t;
-typedef signed short int __int16_t;
-typedef unsigned short int __uint16_t;
-typedef signed int __int32_t;
-typedef unsigned int __uint32_t;
-
-typedef signed long int __int64_t;
-typedef unsigned long int __uint64_t;
-
-
-
-
-
-
-typedef __int8_t __int_least8_t;
-typedef __uint8_t __uint_least8_t;
-typedef __int16_t __int_least16_t;
-typedef __uint16_t __uint_least16_t;
-typedef __int32_t __int_least32_t;
-typedef __uint32_t __uint_least32_t;
-typedef __int64_t __int_least64_t;
-typedef __uint64_t __uint_least64_t;
-
-
-
-typedef long int __quad_t;
-typedef unsigned long int __u_quad_t;
-
-
-
-
-
-
-
-typedef long int __intmax_t;
-typedef unsigned long int __uintmax_t;
-# 141 "/usr/include/bits/types.h" 3 4
-# 1 "/usr/include/bits/typesizes.h" 1 3 4
-# 142 "/usr/include/bits/types.h" 2 3 4
-# 1 "/usr/include/bits/time64.h" 1 3 4
-# 143 "/usr/include/bits/types.h" 2 3 4
-
-
-typedef unsigned long int __dev_t;
-typedef unsigned int __uid_t;
-typedef unsigned int __gid_t;
-typedef unsigned long int __ino_t;
-typedef unsigned long int __ino64_t;
-typedef unsigned int __mode_t;
-typedef unsigned long int __nlink_t;
-typedef long int __off_t;
-typedef long int __off64_t;
-typedef int __pid_t;
-typedef struct { int __val[2]; } __fsid_t;
-typedef long int __clock_t;
-typedef unsigned long int __rlim_t;
-typedef unsigned long int __rlim64_t;
-typedef unsigned int __id_t;
-typedef long int __time_t;
-typedef unsigned int __useconds_t;
-typedef long int __suseconds_t;
-typedef long int __suseconds64_t;
-
-typedef int __daddr_t;
-typedef int __key_t;
-
-
-typedef int __clockid_t;
-
-
-typedef void * __timer_t;
-
-
-typedef long int __blksize_t;
-
-
-
-
-typedef long int __blkcnt_t;
-typedef long int __blkcnt64_t;
-
-
-typedef unsigned long int __fsblkcnt_t;
-typedef unsigned long int __fsblkcnt64_t;
-
-
-typedef unsigned long int __fsfilcnt_t;
-typedef unsigned long int __fsfilcnt64_t;
-
-
-typedef long int __fsword_t;
-
-typedef long int __ssize_t;
-
-
-typedef long int __syscall_slong_t;
-
-typedef unsigned long int __syscall_ulong_t;
-
-
-
-typedef __off64_t __loff_t;
-typedef char *__caddr_t;
-
-
-typedef long int __intptr_t;
-
-
-typedef unsigned int __socklen_t;
-
-
-
-
-typedef int __sig_atomic_t;
-# 28 "/usr/include/stdint.h" 2 3 4
-# 1 "/usr/include/bits/wchar.h" 1 3 4
-# 29 "/usr/include/stdint.h" 2 3 4
-# 1 "/usr/include/bits/wordsize.h" 1 3 4
-# 30 "/usr/include/stdint.h" 2 3 4
-
-
-
-
-# 1 "/usr/include/bits/stdint-intn.h" 1 3 4
-# 24 "/usr/include/bits/stdint-intn.h" 3 4
-typedef __int8_t int8_t;
-typedef __int16_t int16_t;
-typedef __int32_t int32_t;
-typedef __int64_t int64_t;
-# 35 "/usr/include/stdint.h" 2 3 4
-
-
-# 1 "/usr/include/bits/stdint-uintn.h" 1 3 4
-# 24 "/usr/include/bits/stdint-uintn.h" 3 4
-typedef __uint8_t uint8_t;
-typedef __uint16_t uint16_t;
-typedef __uint32_t uint32_t;
-typedef __uint64_t uint64_t;
-# 38 "/usr/include/stdint.h" 2 3 4
-
-
-
-# 1 "/usr/include/bits/stdint-least.h" 1 3 4
-# 25 "/usr/include/bits/stdint-least.h" 3 4
-typedef __int_least8_t int_least8_t;
-typedef __int_least16_t int_least16_t;
-typedef __int_least32_t int_least32_t;
-typedef __int_least64_t int_least64_t;
-
-
-typedef __uint_least8_t uint_least8_t;
-typedef __uint_least16_t uint_least16_t;
-typedef __uint_least32_t uint_least32_t;
-typedef __uint_least64_t uint_least64_t;
-# 42 "/usr/include/stdint.h" 2 3 4
-
-
-
-
-
-typedef signed char int_fast8_t;
-
-typedef long int int_fast16_t;
-typedef long int int_fast32_t;
-typedef long int int_fast64_t;
-# 60 "/usr/include/stdint.h" 3 4
-typedef unsigned char uint_fast8_t;
-
-typedef unsigned long int uint_fast16_t;
-typedef unsigned long int uint_fast32_t;
-typedef unsigned long int uint_fast64_t;
-# 76 "/usr/include/stdint.h" 3 4
-typedef long int intptr_t;
-
-
-typedef unsigned long int uintptr_t;
-# 90 "/usr/include/stdint.h" 3 4
-typedef __intmax_t intmax_t;
-typedef __uintmax_t uintmax_t;
-# 53 "/usr/lib/clang/18/include/stdint.h" 2 3
-# 6 "font.h" 2
-# 25 "font.h"
-const uint64_t FONT_A =
-     0b0011110001100110011001100111111011000011110000111100001111000011;
-# 35 "font.h"
-const uint64_t FONT_B =
-     0b1111110011000011110000111111110011111100110000111100001111111100;
-# 45 "font.h"
-const uint64_t FONT_C =
-     0b0011111111000000110000001100000011000000110000001100000000111111;
-# 55 "font.h"
-const uint64_t FONT_D =
-     0b1111000011001100110000111100001111000011110000111100110011110000;
-# 65 "font.h"
-const uint64_t FONT_E =
-     0b1111111111111111110000001111111111111111110000001111111111111111;
-# 75 "font.h"
-const uint64_t FONT_F =
-     0b1111111111111111110000001111111111111111110000001100000011000000;
-# 85 "font.h"
-const uint64_t FONT_G =
-     0b0111111111111111110000001101111011011111110000111100001111111110;
-# 96 "font.h"
-const uint64_t FONT_H =
-     0b1100001111000011110000111111111111111111110000111100001111000011;
-# 106 "font.h"
-const uint64_t FONT_I =
-     0b1111111111111111000110000001100000011000000110001111111111111111;
-# 116 "font.h"
-const uint64_t FONT_J =
-     0b0000001100000011000000110000001100000011000000110000001111111100;
-# 126 "font.h"
-const uint64_t FONT_K =
-     0b1100011111001100110110001111000011110000110110001100110011000111;
-# 136 "font.h"
-const uint64_t FONT_L =
-     0b1100000011000000110000001100000011000000110000001100000011111111;
-# 146 "font.h"
-const uint64_t FONT_M =
-     0b1100001111100111111111111101101111011011110000111100001111000011;
-# 156 "font.h"
-const uint64_t FONT_N =
-     0b1100001111100011111100111111001111011011110110111100111111000111;
-# 166 "font.h"
-const uint64_t FONT_O =
-     0b0111111011111111110000111100001111000011110000111111111101111110;
-# 176 "font.h"
-const uint64_t FONT_P =
-     0b1111111011111111110000111100001111111111111111101100000011000000;
-# 186 "font.h"
-const uint64_t FONT_Q =
-     0b0111111011111111110000111100001111000011111111110111111000000011;
-# 196 "font.h"
-const uint64_t FONT_R =
-     0b1111111011000011110000111111111011110000110110001100110011000111;
-# 206 "font.h"
-const uint64_t FONT_S =
-     0b0111111111000000110000000011110000000011000000110000001111111110;
-# 216 "font.h"
-const uint64_t FONT_T =
-     0b1111111111111111000110000001100000011000000110000001100000011000;
-# 226 "font.h"
-const uint64_t FONT_U =
-     0b1100001111000011110000111100001111000011110000111100001101111110;
-# 236 "font.h"
-const uint64_t FONT_V =
-     0b1100001111000011110000110110011001100110001111000011110000011000;
-# 246 "font.h"
-const uint64_t FONT_W =
-     0b1100001111000011110000111100001111011011110110111110011111000011;
-# 256 "font.h"
-const uint64_t FONT_X =
-     0b1100001111000011011001100011110000111100011001101100001111000011;
-# 266 "font.h"
-const uint64_t FONT_Y =
-     0b1100001111000011011001100011110000011000000110000001100000011000;
-# 276 "font.h"
-const uint64_t FONT_Z =
-     0b1111111111111111000001110001111001111000111000001111111111111111;
-# 286 "font.h"
-const uint64_t FONT_0 =
-     0b1111111110000001100000011000000110000001100000011000000111111111;
-# 296 "font.h"
-const uint64_t FONT_1 =
-     0b0001100000011000000110000001100000011000000110000001100000011000;
-# 306 "font.h"
-const uint64_t FONT_2 =
-     0b0111111010000011000001100000110000011000001100000110000011111111;
-# 316 "font.h"
-const uint64_t FONT_3 =
-     0b1111110000000010000000010000111000001110000000010000001011111100;
-# 326 "font.h"
-const uint64_t FONT_4 =
-     0b1000000110000001100000011111111100000001000000010000000100000001;
-# 336 "font.h"
-const uint64_t FONT_5 =
-     0b1111111110000000100000001111111100000001000000010000000111111111;
-# 346 "font.h"
-const uint64_t FONT_6 =
-     0b0111110010000000100000001000000010111100100000101000001001111100;
-# 356 "font.h"
-const uint64_t FONT_7 =
-     0b1111111100000001000000100000010000001000000100000010000001000000;
-# 366 "font.h"
-const uint64_t FONT_8 =
-     0b0111111010000001100000010111111010000001100000011000000101111110;
-# 376 "font.h"
-const uint64_t FONT_9 =
-     0b0111111010000001100000010111111100000001000000010000000100000001;
-# 386 "font.h"
-const uint64_t FONT_COLON =
-     0b0000000000011000000110000000000000011000000110000000000000000000;
-# 397 "font.h"
-const uint64_t FONT_CHAR_TABLE[256] = {
-
-    ['A'] = FONT_A,
-    ['B'] = FONT_B,
-    ['C'] = FONT_C,
-    ['D'] = FONT_D,
-    ['E'] = FONT_E,
-    ['F'] = FONT_F,
-    ['G'] = FONT_G,
-    ['H'] = FONT_H,
-    ['I'] = FONT_I,
-    ['J'] = FONT_J,
-    ['K'] = FONT_K,
-    ['L'] = FONT_L,
-    ['M'] = FONT_M,
-    ['N'] = FONT_N,
-    ['O'] = FONT_O,
-    ['P'] = FONT_P,
-    ['Q'] = FONT_Q,
-    ['R'] = FONT_R,
-    ['S'] = FONT_S,
-    ['T'] = FONT_T,
-    ['U'] = FONT_U,
-    ['V'] = FONT_V,
-    ['W'] = FONT_W,
-    ['X'] = FONT_X,
-    ['Y'] = FONT_Y,
-    ['Z'] = FONT_Z,
-    ['0'] = FONT_0,
-    ['1'] = FONT_1,
-    ['2'] = FONT_2,
-    ['3'] = FONT_3,
-    ['4'] = FONT_4,
-    ['5'] = FONT_5,
-    ['6'] = FONT_6,
-    ['7'] = FONT_7,
-    ['8'] = FONT_8,
-    ['9'] = FONT_9,
-    [':'] = FONT_COLON
-};
